requirements.txt
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
dotenv==0.9.9
exceptiongroup==1.3.0
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
huggingface-hub==0.32.3
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
outcome==1.3.0.post0
packaging==25.0
pycparser==2.22
PySocks==1.7.1
python-dotenv==1.1.0
python-telegram-bot==20.7
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
selenium==4.33.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sympy==1.14.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.52.4
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==2.4.0
websocket-client==1.8.0
wsproto==1.2.0
